fix: add rustitude.minimizer to pyi so that it shows up in LSP results
